+++ Changelog.md
@@ -4,6 +4,9 @@
 
 ### Ignition Rendering 2.0.0 (20XX-XX-XX)
 
+1. Move scale functions from Visual to Node class.
+    * [Pull request 159](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/159)
+
 1. Add support for Render Passes and implement Gaussian Noise pass
     * [Pull request 154](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/154)
     * [Pull request 155](https://bitbucket.org/ignitionrobotics/ign-rendering/pull-requests/155)